[NFR] Data access (#11)
* move saveChanges() to separate method in repositories

* remove nullable type

* introduce unitOfWork

* inject unitOfWork instead of repositories

* rename UpdateAsync -> Update

* refactoring services/repos to use with unitOfWork

* fixing coupon precision

+++ src/Mantasflowers.Contracts/Payment/Request/PostCreateCouponRequest.cs
@@ -14,8 +14,8 @@ namespace Mantasflowers.Contracts.Payment.Request
 
         public DateTime RedeemBy { get; set; }
 
-        public long DiscountPrice { get; set; }
+        public decimal DiscountPrice { get; set; }
 
-        public long OrderOverPrice { get; set; }
+        public decimal OrderOverPrice { get; set; }
     }
 }
